refactor: update the path of logger package
