+++ core/quic/quic-go.go
@@ -16,7 +16,7 @@ import (
 	"time"
 
 	quicGo "github.com/lucas-clemente/quic-go"
-	"github.com/yomorun/yomo/logger"
+	"github.com/yomorun/yomo/pkg/logger"
 )
 
 type quicGoServer struct {
