+++ zipper/dispatch.go
@@ -8,7 +8,7 @@ import (
 	"github.com/yomorun/yomo/core/quic"
 	"github.com/yomorun/yomo/internal/core"
 	"github.com/yomorun/yomo/internal/frame"
-	"github.com/yomorun/yomo/logger"
+	"github.com/yomorun/yomo/pkg/logger"
 	"github.com/yomorun/yomo/zipper/tracing"
 )
